moved message factory
