+++ lib_financial_exchange/message_factory/__init__.py
@@ -1,0 +1,2 @@
+
+from lib_financial_exchange.message_factory.message_factory import MessageFactory
